// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const AccountRole = {
  ADMIN: 'ADMIN',
  FARMER: 'FARMER',
  TRANSPORTER: 'TRANSPORTER',
  WAREHOUSE: 'WAREHOUSE',
  STORE: 'STORE',
  INSPECTOR: 'INSPECTOR'
} as const

export type AccountRole = (typeof AccountRole)[keyof typeof AccountRole]


export const AccountStatus = {
  ACTIVE: 'ACTIVE',
  LOCKED: 'LOCKED'
} as const

export type AccountStatus = (typeof AccountStatus)[keyof typeof AccountStatus]


export const PartnerType = {
  FARM: 'FARM',
  TRANSPORT_COMPANY: 'TRANSPORT_COMPANY',
  STORE: 'STORE'
} as const

export type PartnerType = (typeof PartnerType)[keyof typeof PartnerType]


export const CooperationStatus = {
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  TERMINATED: 'TERMINATED'
} as const

export type CooperationStatus = (typeof CooperationStatus)[keyof typeof CooperationStatus]


export const BatchStatus = {
  CREATED: 'CREATED',
  AT_WAREHOUSE: 'AT_WAREHOUSE',
  IN_TRANSIT: 'IN_TRANSIT',
  DELIVERED: 'DELIVERED',
  CANCELLED: 'CANCELLED'
} as const

export type BatchStatus = (typeof BatchStatus)[keyof typeof BatchStatus]


export const CertificateType = {
  VietGAP: 'VietGAP',
  GlobalGAP: 'GlobalGAP',
  HACCP: 'HACCP',
  OTHER: 'OTHER'
} as const

export type CertificateType = (typeof CertificateType)[keyof typeof CertificateType]


export const CertificateStatus = {
  ACTIVE: 'ACTIVE',
  EXPIRED: 'EXPIRED',
  REVOKED: 'REVOKED'
} as const

export type CertificateStatus = (typeof CertificateStatus)[keyof typeof CertificateStatus]


export const TransportStatus = {
  PENDING_PICKUP: 'PENDING_PICKUP',
  IN_TRANSIT: 'IN_TRANSIT',
  ARRIVED_WAREHOUSE: 'ARRIVED_WAREHOUSE',
  DELIVERED: 'DELIVERED'
} as const

export type TransportStatus = (typeof TransportStatus)[keyof typeof TransportStatus]


export const IncidentType = {
  DAMAGED: 'DAMAGED',
  MISSING: 'MISSING',
  QUALITY_ISSUE: 'QUALITY_ISSUE',
  OTHER: 'OTHER'
} as const

export type IncidentType = (typeof IncidentType)[keyof typeof IncidentType]


export const QrScanResult = {
  SUCCESS: 'SUCCESS',
  INACTIVE: 'INACTIVE',
  NOT_FOUND: 'NOT_FOUND'
} as const

export type QrScanResult = (typeof QrScanResult)[keyof typeof QrScanResult]


export const QrCodeStatus = {
  ACTIVE: 'ACTIVE',
  INACTIVE: 'INACTIVE'
} as const

export type QrCodeStatus = (typeof QrCodeStatus)[keyof typeof QrCodeStatus]